Add files via upload
Sentiment analysis provides an in-depth view of the user opinion and helps the organization develop and modify the services for user satisfaction. There is a need for automated opinion classification in every organization related to end user service. Airline review classification, tweet classification, etc. are some of the popular examples of sentiment analysis that are used to improve services and provides a base to add more relevant features. Similarly, the process of sentiment analysis can be applied to reviews of musical instruments by amazon users in a way that would be helpful for an organization serve better to their customers based on their customer reviews. To achieve this objective this project aims to build an opinion collector system for reviews on musical instruments bought by Amazon users.